♻️ refactor(entities): update data types for unitWeight and performanceValue fields
- Changed `unitWeight` from `Int` to `Double` in `MaizeMarket` and `PotatoMarket` entities.
- Updated `performanceValue` default from `String?` to non-null `String` with a default value of "3".

Improves data consistency and ensures correct type usage. No functional changes introduced.

+++ app/src/main/java/com/akilimo/mobile/entities/MaizeMarket.kt
@@ -31,7 +31,7 @@ data class MaizeMarket(
     var unitPrice: Double = 0.0,
 
     @ColumnInfo(name = "unit_weight")
-    var unitWeight: Int = 0,
+    var unitWeight: Double = 0.0,
 
     @ColumnInfo(name = "produce_type")
     var produceType: String = EnumMaizeProduceType.GRAIN.name,
